requirements.txt
requests~=2.28.2
twilio~=8.11.1
python-dotenv~=1.0.0